Add configuration options, modularize, +testing
Working to modularize the application so that it is configurable,
and will once again execute GPIO commands. This will not be
published to NPM until complete, however it is fairly close. This
should be much more useful than the initial single-script application,
and could potentially be useful to other developers, although I do
see this as somewhat unlikely.